+++ src/settings/MappingProfiles/detailsSections/view/ItemDetailSection/AdministrativeData.js
@@ -73,7 +73,7 @@ export const AdministrativeData = ({ mappingDetails }) => {
 
   return (
     <Accordion
-      id="administrative-data"
+      id="view-administrative-data"
       label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.administrativeData.section`} />}
     >
       <Row left="xs">
@@ -128,11 +128,11 @@ export const AdministrativeData = ({ mappingDetails }) => {
       <Row left="xs">
         <Col
           data-test-former-ids
-          id="section-former-ids"
           xs={12}
           className={css.colWithTable}
         >
           <ViewRepeatableField
+            columnIdPrefix="former-ids"
             repeatableAction={formerIdsRepeatableAction}
             fieldData={formerIdentifiersData}
             visibleColumns={formerIdentifiersVisibleColumns}
@@ -145,10 +145,10 @@ export const AdministrativeData = ({ mappingDetails }) => {
       <Row left="xs">
         <Col
           data-test-statistical-codes
-          id="section-statistical-code-ids"
           xs={12}
         >
           <ViewRepeatableField
+            columnIdPrefix="statistical-codes"
             repeatableAction={statisticalCodesRepeatableAction}
             fieldData={statisticalCodesData}
             visibleColumns={statisticalCodesVisibleColumns}